Helix integration working and all the preliminary issues sorted out.
There are a bunch of undecided and untested things, but what is tested is the specialized flow for making reforge meshes from 2-dimensional splines. There is still a lot of work for axes creation and presentation for it to be truly effective.

Lots of refactors and work need to be done too to be honest going forward. But this unblocks a huge amount of the work.